🐞 fix: hotfix expo crash
When example is run with expo, it crashes due to incomplete typecript declaration files (`.d.ts`)  import.

For e.g we do `import T from 'types'` instead of `import T from 'types.d'`. This works in non-expo projects, but crashes the app when run with expo.

This commit fixes the issue